Cn now load fileDB through a seedGenerator file
