+++ index.js
@@ -1,16 +1,19 @@
+require('dotenv').config();
 const express = require('express');
 const app = express();
 const moviesRoutes = require('./routes/movies');
+const { seed } = require('./test/seedGenerator');
 
+seed();
+
+const PORT = process.env.NODE_ENV === 'test' ? 8001 : 8000;
 
 app.use(express.json());
-
 app.use('/v1/movies', moviesRoutes);
 
+const server = app.listen(PORT, () => {
+  console.log(`Listening on port ${PORT}`);
+})
 
-// app.listen('8000', () => {
-//   console.log('Listening on port 8000');
-// })
 
-
-module.exports = app;
+module.exports = { app, server };
