+++ jest.config.js
@@ -1,0 +1,5 @@
+require('dotenv').config({ path: '.env.test' });
+
+module.exports = {
+  // other Jest config
+};
