+++ prod/comments.json
@@ -1,0 +1,1 @@
+{"response":[{"id":1,"movie_id":3,"author":"movie critic","content":"Mind-bending action movie.","created_at":"2024-08-12T15:11:27.765Z"},{"id":2,"movie_id":1,"author":"Fantastic man","content":"A thought-provoking sci-fi movie.","created_at":"2024-08-12T15:11:27.765Z"},{"id":3,"movie_id":8,"author":"uo and down","content":"I need more movies like this.","created_at":"2024-08-12T15:11:27.765Z"},{"id":4,"movie_id":10,"author":"action lamb","content":"Next part please.","created_at":"2024-08-12T15:11:27.765Z"},{"id":5,"movie_id":5,"author":"common man","content":"This makes it my 10th watch today.","created_at":"2024-08-12T15:11:27.765Z"},{"id":6,"movie_id":3,"author":"InceptionFan2","content":"I hate movies like this.","created_at":"2024-10-12T15:11:27.765Z"},{"id":7,"movie_id":10,"author":"InceptionFan2","content":"Too much action.","created_at":"2024-15-12T15:11:27.765Z"},{"id":8,"movie_id":1,"author":"Fan boy","content":"This makes it my 100th watch today.","created_at":"2024-01-12T15:11:27.765Z"},{"id":9,"movie_id":10,"author":"Miami","content":"Do a cartoon version please...","created_at":"2024-25-12T15:15:27.765Z"},{"id":10,"movie_id":10,"author":"Loona","content":"The best this month so far.","created_at":"2024-15-12T15:19:00.765Z"}]}
